restructure the app
