+++ sun/index.html
@@ -13,9 +13,7 @@
     </style>
   </head>
   <body></body>
-  <script src="../suncalc.js"></script>
-  <script src="../js/three.min.js"></script>
-  <script src="../js/OrbitControls.js"></script>
+  <script src="../../suncalc.js"></script>
 
-  <script type="module" src="index.js"></script>
+  <script type="module" src="./js/index.js"></script>
 </html>
